pyproject.toml
[project]
name = "wmata-etl"
version = "1.0.0"
description = "Production-grade ETL pipeline for WMATA rail predictions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Data Engineering Team"}
]
keywords = ["etl", "wmata", "airflow", "data-engineering", "postgresql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "sqlalchemy>=1.4.36,<2.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.9.0",
]
dashboard = [
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
]
dbt = [
    "dbt-postgres>=1.7.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.0.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.0.0",
]
all = [
    "wmata-etl[airflow,dashboard,dbt,dev]",
]

[project.urls]
Repository = "https://github.com/DelphinKdl/metro-transit-etl-pipeline"
Documentation = "https://github.com/DelphinKdl/metro-transit-etl-pipeline#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
